requirements.txt
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
pytz==2017.2
